Predictions and actions
Remove predict_groupR function, replacing it with predict_plot function, which works for both grouped and ungrouped models. The colors were removed from the predict_plot function, as they caused a lot of issues. Users are expected to assign their own colors to prediction plots post-generation. (scico is still recommended for this). Also added a simple function to calculate the pseudo R^2 for a glm, and a Github Action to automatically test the package on different R versions (4.1.1 and 4.3.3 for now)